Cleanup code.

+++ source/vk_descriptors.cpp
@@ -153,7 +153,7 @@ void DescriptorAllocatorGrowable::init(VkDevice device, uint32_t maxSets, std::s
 	
     VkDescriptorPool newPool = createPool(device, maxSets, poolRatios);
 
-    setsPerPool = maxSets * 1.5; //grow it next allocation
+    setsPerPool = uint32_t(maxSets * 1.5f); //grow it next allocation
 
     readyPools.push_back(newPool);
 }
@@ -195,7 +195,7 @@ VkDescriptorPool DescriptorAllocatorGrowable::getPool(VkDevice device)
 	    //need to create a new pool
 	    newPool = createPool(device, setsPerPool, ratios);
 
-	    setsPerPool = setsPerPool * 1.5;
+	    setsPerPool = uint32_t(setsPerPool * 1.5);
 	    if (setsPerPool > 4092) {
 		    setsPerPool = 4092;
 	    }
